+++ tests/cypress/support/commands_projects.js
@@ -28,9 +28,9 @@ Cypress.Commands.add(
         cy.get('.cvat-create-project-content').within(() => {
             cy.contains('Submit').click();
         });
-        if (expectedResult == 'success') {
+        if (expectedResult === 'success') {
             cy.get('.cvat-notification-create-project-success').should('exist').find('[data-icon="close"]').click();
-        } else if (expectedResult == 'fail') {
+        } else if (expectedResult === 'fail') {
             cy.get('.cvat-notification-create-project-success').should('not.exist');
         }
         cy.goToProjectsList();
@@ -67,9 +67,10 @@ Cypress.Commands.add('deleteProject', (projectName, projectID, expectedResult = 
     }
 });
 
-Cypress.Commands.add('exportProject', ({ projectName, as, type, dumpType, archiveCustomeName }) => {
+Cypress.Commands.add('exportProject', ({
+    projectName, type, dumpType, archiveCustomeName,
+}) => {
     cy.projectActions(projectName);
-    cy.intercept('GET', `/api/v1/projects/**/${type}**`).as(as);
     cy.get('.cvat-project-actions-menu').contains('Export project dataset').click();
     cy.get('.cvat-modal-export-project').should('be.visible').find('.cvat-modal-export-select').click();
     cy.contains('.cvat-modal-export-option-item', dumpType).should('be.visible').click();
@@ -82,8 +83,20 @@ Cypress.Commands.add('exportProject', ({ projectName, as, type, dumpType, archiv
     }
     cy.get('.cvat-modal-export-project').contains('button', 'OK').click();
     cy.get('.cvat-notification-notice-export-project-start').should('be.visible');
-    cy.wait(`@${as}`, { timeout: 5000 }).its('response.statusCode').should('equal', 202);
-    cy.wait(`@${as}`).its('response.statusCode').should('equal', 201);
+});
+
+Cypress.Commands.add('getDownloadFileName', () => {
+    cy.intercept('GET', '**=download').as('download');
+    cy.wait('@download').then((download) => {
+        const filename = download.response.headers['content-disposition'].split('filename="b\'')[1].split('\'')[0];
+        return filename;
+    });
+});
+
+Cypress.Commands.add('waitForDownload', () => {
+    cy.getDownloadFileName().then((filename) => {
+        cy.verifyDownload(filename);
+    });
 });
 
 Cypress.Commands.add('deleteProjectViaActions', (projectName) => {